fix review issues: merge conflicts, N+1 queries, CORS preflight, error handling
- resolve merge conflicts with main (enums re-export, migration ID collisions)
- add AuditAction to core/enums.py, convert models/enums.py to re-exports
- fix migration duplicate revision IDs (a1b2c3d4e5f6 -> 3edc932f71fd, a2b3c4d5e6f7 -> 8e1cc0628ef4)
- replace 4N+1 loop queries in list_airports_admin with subquery aggregates
- add 30s TTL cache for maintenance mode DB check to reduce connection churn
- allow OPTIONS preflight through maintenance middleware for proper CORS
- fix bare catch in InviteUserDialog to capture actual error message
